requirements.txt
Flask
pymongo
python-dotenv
Flask-Bcrypt
Flask-JWT-Extended
